# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API De Tuna Resort",
        "description": "Dokumentasi API untuk Sistem Informasi & Pemesanan De Tuna Resort",
        "contact": {
            "email": "admin@detuna.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server API De Tuna (Lokal - php artisan serve)"
        },
        {
            "url": "http://detuna-api.test",
            "description": "Server API De Tuna (Lokal - Laragon)"
        }
    ],
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi Pengguna Baru",
                "operationId": "ad77fe138bcb384ad65d167d5c8d1cc0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "08123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login Pengguna",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|Abc..."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Kredensial salah"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout Pengguna",
                "operationId": "c5cdd87fdbf9d0f95dede875bd023dec",
                "responses": {
                    "200": {
                        "description": "Logout berhasil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get Data Pengguna (Current)",
                "operationId": "13b43df400e475a157538f73617b8c3d",
                "responses": {
                    "200": {
                        "description": "Data pengguna",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bookings": {
            "post": {
                "tags": [
                    "Bookings (Pelanggan)"
                ],
                "summary": "Membuat booking baru (Pelanggan)",
                "operationId": "9ec8aa3de0d1efdd0cfe72c27eac745c",
                "requestBody": {
                    "description": "Data booking baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Token tidak valid)"
                    },
                    "422": {
                        "description": "Validasi gagal (Data input tidak valid)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/my-bookings": {
            "get": {
                "tags": [
                    "Bookings (Pelanggan)"
                ],
                "summary": "Melihat riwayat booking (Pelanggan)",
                "operationId": "52ff3343d858b466349c0b70c55476aa",
                "responses": {
                    "200": {
                        "description": "Daftar riwayat booking",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Booking"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Token tidak valid)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/rooms/{id}/reviews": {
            "get": {
                "tags": [
                    "Reviews (Pelanggan & Publik)"
                ],
                "summary": "Get semua ulasan untuk satu kamar (Publik)",
                "operationId": "a7da69fd8c0de5172939975860676d0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "Room ID",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar ulasan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "user_name": {
                                                "type": "string",
                                                "example": "Budi S."
                                            },
                                            "rating": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "comment": {
                                                "type": "string",
                                                "example": "Keren!"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kamar tidak ditemukan"
                    }
                }
            }
        },
        "/api/v1/reviews": {
            "post": {
                "tags": [
                    "Reviews (Pelanggan & Publik)"
                ],
                "summary": "Membuat ulasan baru (Pelanggan)",
                "operationId": "20f4daafc07903e8535348e0047f7920",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ulasan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreBookingRequest": {
                "title": "Store Booking Request",
                "required": [
                    "room_id",
                    "check_in_date",
                    "check_out_date",
                    "total_price"
                ],
                "properties": {
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "check_in_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-20"
                    },
                    "check_out_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-22"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "decimal",
                        "example": 5000000
                    },
                    "guests": {
                        "description": "Jumlah tamu",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "StoreReviewRequest": {
                "title": "Store Review Request",
                "required": [
                    "booking_id",
                    "rating"
                ],
                "properties": {
                    "booking_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "rating": {
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 5
                    },
                    "comment": {
                        "type": "string",
                        "example": "Sangat memuaskan!"
                    }
                },
                "type": "object"
            },
...
            "Booking": {
                "title": "Booking",
                "description": "Model Pemesanan",
                "properties": {
                    "booking_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "booking_code": {
                        "type": "string",
                        "example": "DT-20251021-XYZ"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "check_in_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-20"
                    },
                    "check_out_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-22"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "decimal",
                        "example": 5000000
                    },
                    "booking_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "canceled",
                            "completed"
                        ],
                        "example": "pending"
                    },
                    "payment_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "paid",
                            "failed"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Review": {
                "title": "Review",
                "description": "Model Ulasan Pelanggan",
                "properties": {
                    "review_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "booking_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "rating": {
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 5
                    },
                    "comment": {
                        "type": "string",
                        "example": "Pelayanan sangat memuaskan!"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Model Pengguna",
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "full_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "pelanggan",
                            "admin",
                            "super_admin"
                        ],
                        "example": "pelanggan"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token (Bearer Token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "sanctum": {
            "type": "http",
            "description": "Masukkan Bearer Token (Sanctum)",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}